feat(initial): initial commit
This is the initial version of the code for the project. It can
successfully publish to MQTT topics - both to random topics in a simple
mode and also so Home Assistant can discover the stuff, but there are
still more things that can be added to the HASS integration - such as
specifying the icons and modularizing the HASS code even more (it
already is better than my other projects). However, before I do that,
it'll get the ability to write results to InfluxDB so it stays
consistent with my other projects.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>