fix(datatypes): change firmware to a tag
Firmware is really something that should be a tag so you can filter more
easily based on versions of the software. It's also something that
doesn't change that often. Hopefully no one grabbed this before I made
this change, because it's a pain to cleanup.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>